third_party/WebKit/LayoutTests/fast/forms/file/input-file-element-clone.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test.js"></script>
<script src="resources/file-drag-common.js"></script>
<form  method="GET">
<input type="file" id="file1">
</form>
<script>
description('Make sure that Node.cloneNode() works as expected for file input type.');
var file1 = document.getElementById('file1');
dragFilesOntoInput(file1, ['input-file-element-clone.html']);
var clone = file1.cloneNode(true);
shouldBeEqualToString("file1.value", "C:\\fakepath\\input-file-element-clone.html");
clone.id = "file2";
document.body.appendChild(clone);
var file2 = document.getElementById('file2');
shouldBeEqualToString("file2.value", "C:\\fakepath\\input-file-element-clone.html");
dragFilesOntoInput(file1, ['input-file-element-clone-expected.txt']);
shouldBeEqualToString("file1.value", "C:\\fakepath\\input-file-element-clone-expected.txt");
shouldBeEqualToString("file2.value", "C:\\fakepath\\input-file-element-clone.html");
</script>
</body>
</html>
