third_party/WebKit/LayoutTests/fast/forms/file/input-file-element-clone-expected.txt

Make sure that Node.cloneNode() works as expected for file input type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS file1.value is "C:\\fakepath\\input-file-element-clone.html"
PASS file2.value is "C:\\fakepath\\input-file-element-clone.html"
PASS file1.value is "C:\\fakepath\\input-file-element-clone-expected.txt"
PASS file2.value is "C:\\fakepath\\input-file-element-clone.html"
PASS successfullyParsed is true

TEST COMPLETE
